requirements.txt
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
yt-dlp>=2024.1.0
